backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
pytest==8.4.1
httpx==0.28.1
python-multipart==0.0.20
pypdf==6.1.1
